Refactor and fix test warning for syncStarknetState
Was initially just going to fix the warning in the test case, but decided to refactor the file a bit so we can properly mock out calls to IPFS in the tests.
